Move GetAnime.js to components folder
